comment and reorg
