+++ source/GameInterface/Services/Armies/Extensions/ArmyExtensions.cs
@@ -2,46 +2,56 @@
 using Common.Logging;
 using GameInterface.Services.ObjectManager;
 using Serilog;
-using Serilog.Core;
 using System;
 using System.Reflection;
 using TaleWorlds.CampaignSystem;
+using TaleWorlds.CampaignSystem.Actions;
 using TaleWorlds.CampaignSystem.Party;
 
-namespace GameInterface.Services.Armies.Extensions
+namespace GameInterface.Services.Armies.Extensions;
+
+/// <summary>
+/// Extensions for the <see cref="Army"/> class
+/// </summary>
+internal static class ArmyExtensions
 {
-    internal static class ArmyExtensions
+    private static readonly ILogger Logger = LogManager.GetLogger<Army>();
+
+    private static Action<Army, MobileParty> Army_OnAddPartyInternal = typeof(Army).GetMethod("OnAddPartyInternal", BindingFlags.NonPublic | BindingFlags.Instance)
+        .BuildDelegate<Action<Army, MobileParty>>();
+    private static Action<Army, MobileParty> Army_OnRemovePartyInternal = typeof(Army).GetMethod("OnRemovePartyInternal", BindingFlags.NonPublic | BindingFlags.Instance)
+        .BuildDelegate<Action<Army, MobileParty>>();
+    private static Action<Army, Army.ArmyDispersionReason> DisbandArmyAction_ApplyInternal = typeof(DisbandArmyAction)
+            .GetMethod("ApplyInternal", BindingFlags.NonPublic | BindingFlags.Static)
+            .BuildDelegate<Action<Army, Army.ArmyDispersionReason>>();
+
+    internal static void AddPartyInternal(this MobileParty mobileParty, Army army)
     {
-        private static readonly ILogger Logger = LogManager.GetLogger<Army>();
+        Army_OnAddPartyInternal(army,mobileParty);
+    }
+    internal static void RemovePartyInternal(this MobileParty mobileParty, Army army)
+    {
+        Army_OnRemovePartyInternal(army,mobileParty);
+    }
 
-        private static Action<Army, MobileParty> Army_OnAddPartyInternal = typeof(Army).GetMethod("OnAddPartyInternal", BindingFlags.NonPublic | BindingFlags.Instance)
-            .BuildDelegate<Action<Army, MobileParty>>();
-        private static Action<Army, MobileParty> Army_OnRemovePartyInternal = typeof(Army).GetMethod("OnRemovePartyInternal", BindingFlags.NonPublic | BindingFlags.Instance)
-            .BuildDelegate<Action<Army, MobileParty>>();
-        
-        internal static void OnAddPartyInternal(this MobileParty mobileParty, Army army)
+    internal static void DisbandArmy(this Army army, Army.ArmyDispersionReason reason)
+    {
+        DisbandArmyAction_ApplyInternal(army, reason);
+    }
+
+    public static string GetStringId(this Army army)
+    {
+        if (ContainerProvider.TryResolve<IObjectManager>(out var objectManager) == false)
         {
-            Army_OnAddPartyInternal(army,mobileParty);
+            Logger.Error("Unable to get {objectManager}", nameof(IObjectManager));
+            return null;
         }
-        internal static void OnRemovePartyInternal(this MobileParty mobileParty, Army army)
+
+        if (objectManager.TryGetId(army, out var id) == false)
         {
-            Army_OnRemovePartyInternal(army,mobileParty);
+            Logger.Error("{army} was not properly registered", army.Name);
         }
 
-        public static string GetStringId(this Army army)
-        {
-            if (ContainerProvider.TryResolve<IObjectManager>(out var objectManager) == false)
-            {
-                Logger.Error("Unable to get {objectManager}", nameof(IObjectManager));
-                return null;
-            }
-
-            if (objectManager.TryGetId(army, out var id) == false)
-            {
-                Logger.Error("{army} was not properly registered", army.Name);
-            }
-
-            return id;
-        }
+        return id;
     }
 }
