--- source/GameInterface/Services/Armies/Messages/DisbandArmy.cs
@@ -1,15 +1,0 @@
-﻿using Common.Messaging;
-using GameInterface.Services.Armies.Data;
-
-namespace GameInterface.Services.Armies.Messages
-{
-    public record DisbandArmy : ICommand
-    {
-        public ArmyDeletionData Data { get; }
-
-        public DisbandArmy(ArmyDeletionData armyDeletionData)
-        {
-            Data = armyDeletionData;
-        }
-    }
-}
