+++ source/GameInterface/Services/Armies/Messages/MobilePartyInArmyAdded.cs
@@ -1,17 +1,19 @@
 ﻿using Common.Messaging;
 
-namespace GameInterface.Services.Armies.Messages
+namespace GameInterface.Services.Armies.Messages;
+
+/// <summary>
+/// Event for when a MobileParty is added to an Army
+/// </summary>
+public record MobilePartyInArmyAdded : IEvent
 {
-    public record MobilePartyInArmyAdded : ICommand
+    public string MobilePartyId { get; }
+    public string LeaderMobilePartyId { get; }
+
+    public MobilePartyInArmyAdded(string mobilePartyId, string leaderMobilePartyId)
     {
-        public string MobilePartyId { get; }
-        public string LeaderMobilePartyId { get; }
-
-        public MobilePartyInArmyAdded(string mobilePartyId, string leaderMobilePartyId)
-        {
-            MobilePartyId = mobilePartyId;
-            LeaderMobilePartyId = leaderMobilePartyId;
-        }
-
+        MobilePartyId = mobilePartyId;
+        LeaderMobilePartyId = leaderMobilePartyId;
     }
+
 }
